Implementations of Friedman test
This script does different variations on the Friedman test. The
motivation is that other implementations seem to use the Chi-square
distribution approximation, which (supposedly) is poor for small sample
sizes.

This iteration includes a version of Friedman test referenced in
Conover, Practical Nonparameteric Statistics, 3rd Ed., and a bootstrap
version. In the future, I want to add an exact version for small sample
sizes, and multiple comparisons procedures (Conover's recommended
method).